examples/unstructured-to-structured/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
openai>=1.40.0
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.1

